requirements.txt
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.0
idna==2.8
nose==1.3.7
py==1.8.0
retry==0.9.2
selenium==3.141.0
six==1.12.0
spec==1.4.1
urllib3==1.25.3
Appium-Python-Client==0.47

facebook-wda~=1.4.7
numpy~=1.26.4
pillow~=10.3.0
requests~=2.31.0